rename consts to navitems
