+++ src/components/Navbar/Mobile/MobileNavigation.tsx
@@ -6,8 +6,8 @@ import { ThemeContext } from 'contexts/themeContext'
 import { useWallet } from 'hooks/Wallet'
 import Link from 'next/link'
 import { useContext, useEffect, useState } from 'react'
-import { mobileNavItems } from '../constants'
 import Logo from '../Logo'
+import { mobileNavItems } from '../navigationItems'
 import { topNavStyles } from '../navStyles'
 import { TransactionsList } from '../TransactionList'
