+++ src/components/Navbar/SiteNavigation.tsx
@@ -3,8 +3,8 @@ import { Header } from 'antd/lib/layout/layout'
 import useMobile from 'hooks/Mobile'
 import { CSSProperties, useEffect, useState } from 'react'
 import Account from './Account'
-import { desktopMenuItems, resourcesMenuItems } from './constants'
 import MobileNavigation from './Mobile/MobileNavigation'
+import { desktopMenuItems, resourcesMenuItems } from './navigationItems'
 import NavLanguageSelector from './NavLanguageSelector'
 import { topNavStyles } from './navStyles'
 import ThemePicker from './ThemePicker'
